refactor(heuristics): extract target-generator into its own module
Move TargetContext, TargetGeneratorABC, DefaultTargetGenerator,
_validate_candidate, TargetGeneratorCallable, _CallableTargetGenerator,
and _coerce_target_generator out of movement.py into a dedicated
physical/target_generator.py. movement.py now imports from the new
module and keeps its focus on traversal orchestration.

Also delete the now-unused _target_from_stage_controls_only helper on
PhysicalPlacementStrategy — its behavior is fully covered by
DefaultTargetGenerator.generate.

No API change: physical/__init__.py continues to re-export the public
symbols (TargetContext, TargetGeneratorABC, DefaultTargetGenerator);
they are just imported from the new location.
